# bcs-services/cluster-resources/proto/cluster-resources/cluster-resources.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cluster Resources ApiDoc",
    "version": "1.0",
    "license": {
      "name": "MIT"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/clusterresources/v1/projects/{projectID}/clusters/{clusterID}/namespaces/{namespace}/networks/services": {
      "get": {
        "summary": "ListSVC API",
        "description": "获取 Service 列表",
        "operationId": "Network_ListSVC",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/clusterresourcesCommonResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "projectID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "clusterID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "labelSelector",
            "description": "标签选择器.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "apiVersion",
            "description": "apiVersion.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "ownerName",
            "description": "所属资源名称.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "ownerKind",
            "description": "所属资源类型.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "format",
            "description": "资源配置格式（manifest/selectItems）.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "scene",
            "description": "使用场景. 仅 selectItems 格式下有效",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Network"
        ]
      }
    },
    "/clusterresources/v1/projects/{projectID}/clusters/{clusterID}/namespaces/{namespace}/networks/services/{name}": {
      "get": {
        "summary": "GetSVC API",
        "description": "获取 Service",
        "operationId": "Network_GetSVC",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/clusterresourcesCommonResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "projectID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "clusterID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "apiVersion",
            "description": "apiVersion.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "format",
            "description": "资源配置格式（manifest/formData）.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Network"
        ]
      },
      "delete": {
        "summary": "DeleteSVC API",
        "description": "删除 Service",
        "operationId": "Network_DeleteSVC",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/clusterresourcesCommonResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "projectID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "clusterID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Network"
        ]
      },
      "put": {
        "summary": "UpdateSVC API",
        "description": "更新 Service",
        "operationId": "Network_UpdateSVC",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/clusterresourcesCommonResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "projectID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "clusterID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/clusterresourcesResUpdateReq"
            }
          }
        ],
        "tags": [
          "Network"
        ]
      }
    },
...
  },
  "definitions": {
...
    "clusterresourcesCommonResp": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回错误码",
          "title": "code"
        },
        "message": {
          "type": "string",
          "description": "返回错误信息",
          "title": "message"
        },
        "requestID": {
          "type": "string",
          "description": "请求 ID",
          "title": "request id"
        },
        "data": {
          "type": "object",
          "description": "资源信息",
          "title": "data"
        },
        "webAnnotations": {
          "type": "object",
          "description": "web 注解",
          "title": "webAnnotations"
        }
      },
      "description": "通用响应体，适用于资源型 API 请求返回，如需定义详细的 data 结构则不使用该响应体",
      "title": "CommonResp"
    },
...
    "clusterresourcesResUpdateReq": {
      "type": "object",
      "properties": {
        "projectID": {
          "type": "string",
          "title": "项目 ID"
        },
        "clusterID": {
          "type": "string",
          "title": "集群 ID"
        },
        "namespace": {
          "type": "string",
          "title": "命名空间"
        },
        "name": {
          "type": "string",
          "title": "资源名称"
        },
        "rawData": {
          "type": "object",
          "title": "资源配置信息"
        },
        "format": {
          "type": "string",
          "title": "资源配置格式（manifest/formData）"
        }
      },
      "description": "更新单个 k8s 资源请求体",
      "title": "ResUpdateReq"
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\r\nprotocol buffer message. This string must contain at least\r\none \"/\" character. The last segment of the URL's path must represent\r\nthe fully qualified name of the type (as in\r\n`path/google.protobuf.Duration`). The name should be in a canonical form\r\n(e.g., leading \".\" is not accepted).\r\n\r\nIn practice, teams usually precompile into the binary all types that they\r\nexpect it to use in the context of Any. However, for URLs which use the\r\nscheme `http`, `https`, or no scheme, one can optionally set up a type\r\nserver that maps type URLs to message definitions as follows:\r\n\r\n* If no scheme is provided, `https` is assumed.\r\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\r\n  value in binary format, or produce an error.\r\n* Applications are allowed to cache lookup results based on the\r\n  URL, or have them precompiled into a binary to avoid any\r\n  lookup. Therefore, binary compatibility needs to be preserved\r\n  on changes to types. (Use versioned type names to manage\r\n  breaking changes.)\r\n\r\nNote: this functionality is not currently available in the official\r\nprotobuf release, and it is not used for type URLs beginning with\r\ntype.googleapis.com.\r\n\r\nSchemes other than `http`, `https` (or the empty scheme) might be\r\nused with implementation specific semantics."
        },
        "value": {
          "type": "string",
          "format": "byte",
          "description": "Must be a valid serialized protocol buffer of the above specified type."
        }
      },
      "title": "`Any` contains an arbitrary serialized protocol buffer message along with a\r\nURL that describes the type of the serialized message.\r\n\r\nProtobuf library provides support to pack/unpack Any values in the form\r\nof utility functions or additional generated methods of the Any type.\r\n\r\nExample 1: Pack and unpack a message in C++.\r\n\r\n    Foo foo = ...;\r\n    Any any;\r\n    any.PackFrom(foo);\r\n    ...\r\n    if (any.UnpackTo(\u0026foo)) {\r\n      ...\r\n    }\r\n\r\nExample 2: Pack and unpack a message in Java.\r\n\r\n    Foo foo = ...;\r\n    Any any = Any.pack(foo);\r\n    ...\r\n    if (any.is(Foo.class)) {\r\n      foo = any.unpack(Foo.class);\r\n    }\r\n\r\n Example 3: Pack and unpack a message in Python.\r\n\r\n    foo = Foo(...)\r\n    any = Any()\r\n    any.Pack(foo)\r\n    ...\r\n    if any.Is(Foo.DESCRIPTOR):\r\n      any.Unpack(foo)\r\n      ...\r\n\r\n Example 4: Pack and unpack a message in Go\r\n\r\n     foo := \u0026pb.Foo{...}\r\n     any, err := ptypes.MarshalAny(foo)\r\n     ...\r\n     foo := \u0026pb.Foo{}\r\n     if err := ptypes.UnmarshalAny(any, foo); err != nil {\r\n       ...\r\n     }\r\n\r\nThe pack methods provided by protobuf library will by default use\r\n'type.googleapis.com/full.type.name' as the type URL and the unpack\r\nmethods only use the fully qualified type name after the last '/'\r\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\r\nname \"y.z\".\r\n\r\n\r\nJSON\r\n====\r\nThe JSON representation of an `Any` value uses the regular\r\nrepresentation of the deserialized, embedded message, with an\r\nadditional field `@type` which contains the type URL. Example:\r\n\r\n    package google.profile;\r\n    message Person {\r\n      string first_name = 1;\r\n      string last_name = 2;\r\n    }\r\n\r\n    {\r\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\r\n      \"firstName\": \u003cstring\u003e,\r\n      \"lastName\": \u003cstring\u003e\r\n    }\r\n\r\nIf the embedded message type is well-known and has a custom JSON\r\nrepresentation, that representation will be embedded adding a field\r\n`value` which holds the custom JSON in addition to the `@type`\r\nfield. Example (for message [google.protobuf.Duration][]):\r\n\r\n    {\r\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\r\n      \"value\": \"1.212s\"\r\n    }"
    },
...
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
  }
}